requirements.txt
AudioLoader==0.1.0
omegaconf 
tqdm
torch
torchaudio
torchvision
pytorch-lightning
musdb
dora-search
diffq>=0.2.1
hydra-colorlog>=1.1
hydra-core>=1.1
julius>=0.2.3
lameenc>=1.2
museval
mypy
openunmix
pyyaml
